+++ SoomlaAndroidProfile/src/com/soomla/profile/SoomlaProfile.java
@@ -832,7 +832,7 @@ public class SoomlaProfile {
      * @throws ProviderNotFoundException if the supplied provider is not
      *                                   supported by the framework
      */
-    public void submitScore(final IProvider.Provider provider, final Leaderboard leaderboard, final int value, final String payload, final Reward reward) throws ProviderNotFoundException {
+    public void submitScore(final IProvider.Provider provider, final Leaderboard leaderboard, final long value, final String payload, final Reward reward) throws ProviderNotFoundException {
         mGameServicesController.submitScore(provider, leaderboard, value, payload, reward);
     }
